move fragment_correlation, fragment_correlation_different
