+++ tests/unit_tests/test_plexscoring_utils.py
@@ -4,8 +4,11 @@ import pytest
 
 from alphadia.plexscoring.utils import (
     calculate_score_groups,
+    fragment_correlation,
+    fragment_correlation_different,
     merge_missing_columns,
     multiplex_candidates,
+    save_corrcoeff,
 )
 
 
@@ -150,3 +153,65 @@ def test_merge_missing(left_and_right_df):
     df = merge_missing_columns(left_df, right_df, ["col_3"], on="idx")
     # then
     assert np.all(df.columns == ["idx", "col_1", "col_2", "col_3"])
+
+
+def test_save_corrcoeff():
+    p = save_corrcoeff(
+        np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.float32),
+        np.array([10, 9, 8, 7, 6, 5, 4, 3, 2, 1], dtype=np.float32),
+    )
+    assert np.isclose(p, -1.0)
+
+    p = save_corrcoeff(
+        np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.float32),
+        np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.float32),
+    )
+    assert np.isclose(p, 1.0)
+
+    p = save_corrcoeff(
+        np.zeros(10, dtype=np.float32),
+        np.zeros(10, dtype=np.float32),
+    )
+    assert np.isclose(p, 0.0)
+
+
+def test_fragment_correlation():
+    a = np.array(
+        [[[1, 2, 3], [1, 2, 3]], [[3, 2, 1], [1, 2, 3]], [[0, 0, 0], [0, 0, 0]]]
+    )
+    corr = fragment_correlation(a)
+    assert corr.shape == (2, 3, 3)
+
+    test_a = np.array(
+        [
+            [[1.0, -1.0, 0.0], [-1.0, 1.0, 0.0], [0.0, 0.0, 0.0]],
+            [[1.0, 1.0, 0.0], [1.0, 1.0, 0.0], [0.0, 0.0, 0.0]],
+        ]
+    )
+    assert np.allclose(corr, test_a)
+
+    b = np.zeros((10, 10, 10))
+    corr = fragment_correlation(b)
+    assert corr.shape == (10, 10, 10)
+    assert np.allclose(corr, b)
+
+
+def test_fragment_correlation_different():
+    a = np.array(
+        [[[1, 2, 3], [1, 2, 3]], [[3, 2, 1], [1, 2, 3]], [[0, 0, 0], [0, 0, 0]]]
+    )
+    corr = fragment_correlation_different(a, a)
+    assert corr.shape == (2, 3, 3)
+
+    test_a = np.array(
+        [
+            [[1.0, -1.0, 0.0], [-1.0, 1.0, 0.0], [0.0, 0.0, 0.0]],
+            [[1.0, 1.0, 0.0], [1.0, 1.0, 0.0], [0.0, 0.0, 0.0]],
+        ]
+    )
+    assert np.allclose(corr, test_a)
+
+    b = np.zeros((10, 10, 10))
+    corr = fragment_correlation_different(b, b)
+    assert corr.shape == (10, 10, 10)
+    assert np.allclose(corr, b)
